+++ code/cp_assets.cljs
@@ -9,14 +9,6 @@
     {:target-path "resources/public/ui/css/themes"}]
    ["node_modules/react-datepicker/dist/react-datepicker.min.css"
     {:target-path "resources/public/ui/css/react-datepicker.min.css"}]
-   ["node_modules/codemirror/lib/codemirror.css"
-    {:target-path "resources/public/ui/css/codemirror.css"}]
-   ["node_modules/codemirror/addon/fold/foldgutter.css"
-    {:target-path "resources/public/ui/css/foldgutter.css"}]
-   ["node_modules/codemirror/addon/dialog/dialog.css"
-    {:target-path "resources/public/ui/css/dialog.css"}]
-   ["node_modules/codemirror/addon/search/matchesonscrollbar.css"
-    {:target-path "resources/public/ui/css/matchesonscrollbar.css"}]
    ["node_modules/leaflet/dist/leaflet.css"
     {:target-path "resources/public/ui/css/leaflet.css"}]
    ["node_modules/leaflet-draw/dist/leaflet.draw.css"