+++ code/src/cljs/sixsq/nuvla/ui/apps/views.cljs
@@ -33,7 +33,7 @@
 
 
 (defn ModuleDetails
-  [_nav-query-params]
+  []
   (let [module (subscribe [::subs/module])
         new-subtype (subscribe [::route-subs/query-param :subtype])]
     (fn []
@@ -48,19 +48,18 @@
 
 
 (defn Module
-  [_nav-query-params]
+  []
   (dispatch [::main-events/changes-protection? false])
   (dispatch [::events/form-valid true])
   (dispatch [::events/set-validate-form? false])
-  (fn [nav-query-params]
-    [components/LoadingPage {:dimmable? true}
-     [:<>
-      [components/NotFoundPortal
-       ::subs/module-not-found?
-       :no-module-message-header
-       :no-module-message-content]
-      [views-detail/VersionWarning]
-      [ModuleDetails nav-query-params]]]))
+  [components/LoadingPage {:dimmable? true}
+   [:<>
+    [components/NotFoundPortal
+     ::subs/module-not-found?
+     :no-module-message-header
+     :no-module-message-content]
+    [views-detail/VersionWarning]
+    [ModuleDetails]]])
 
 (defn- CommonComponents
   []
@@ -73,18 +72,21 @@
 
 (defn AppDetails
   []
-  (let [nav-query-params (subscribe [::route-subs/nav-query-params])]
+  (let [version      (subscribe [::route-subs/query-param :version])
+        sub-type     (subscribe [::route-subs/query-param :subtype])]
     (fn []
-      (let [version     (:version @nav-query-params)
-            new-subtype (:subtype @nav-query-params)
-            is-new?     (boolean (seq new-subtype))]
+      (let [is-new?  (boolean (seq @sub-type))]
         (dispatch [::events/is-new? is-new?])
         (if is-new?
-          (dispatch-clear-events new-subtype)
-          (dispatch [::events/get-module version]))
+          (dispatch-clear-events @sub-type)
+          (dispatch [::events/get-module @version]))
         [:<>
          [CommonComponents]
-         [Module nav-query-params]]))))
+         [Module]]))))
+
+(defn AppDetailsRoute
+  [{:keys [path-params]}]
+  [AppDetails {:key (str path-params)}])
 
 (defn AppsOverview
   [_path]