+++ code/src/cljs/sixsq/nuvla/ui/main/views.cljs
@@ -116,16 +116,17 @@
 
 (defn ignore-changes-modal
   []
-  (let [tr                (subscribe [::i18n-subs/tr])
-        navigation-info   (subscribe [::subs/ignore-changes-modal])
-        on-apps-page?     (-> @(subscribe [::route-subs/current-route])
-                              :template
-                              (str/includes? "apps"))
-        ignore-changes-fn #(dispatch [::events/ignore-changes false])]
+  (let [tr                       (subscribe [::i18n-subs/tr])
+        navigation-info          (subscribe [::subs/ignore-changes-modal])
+        do-not-ignore-changes-fn #(dispatch [::events/do-not-ignore-changes])]
+    (when @navigation-info
+      (dispatch [::events/opening-protection-modal]))
 
-    [ui/Modal {:open       (some? @navigation-info)
-               :close-icon true
-               :on-close   ignore-changes-fn}
+    [ui/Modal {:open        (some? @navigation-info)
+               :close-icon  true
+               :on-close    do-not-ignore-changes-fn
+                ;; data-testid is used for e2e test
+               :data-testid "protection-modal"}
 
      [uix/ModalHeader {:header (@tr [:ignore-changes?])}]
 
@@ -135,9 +136,8 @@
       [uix/Button {:text     (@tr [:ignore-changes]),
                    :positive true
                    :active   true
-                   :on-click #(do (dispatch [::events/ignore-changes true])
-                                  (dispatch [::apps-events/form-valid])
-                                  (when on-apps-page? (dispatch [::apps-events/set-active-tab :overview])))}]]]))
+                   :on-click #(do (dispatch [::events/ignore-changes])
+                                  (dispatch [::apps-events/form-valid]))}]]]))
 
 (defn subscription-required-modal
   []